fix(dss): resolve compile errors and update unit tests
Fix effectively-final lambda, roundQuantity overload, DssProperties import, priceChangePercent API in tests, and daily sales map mocking.

+++ src/main/java/com/example/secdsp/modules/dss/service/SellerDiscountProfitCalculator.java
@@ -56,7 +56,9 @@ public class SellerDiscountProfitCalculator {
             );
         }
 
-        long forecastDemand = roundQuantity(forecastDemandValue);
+        long forecastDemand = roundQuantity(
+            BigDecimal.valueOf(forecastDemandValue)
+        );
         BigDecimal elasticity = BigDecimal.valueOf(elasticityValue);
         BigDecimal demandGrowthRate = elasticity.multiply(changeRate);
         long predictedDemand = roundQuantity(
